requirements.txt
aiohttp>=3.10.2,<4.0.0
python-dotenv>=0.21.0,<0.22.0
anthropic>=0.5.0,<0.6.0
langdetect>=1.0.9,<2.0.0
python-telegram-bot==20.3